12/05/24/red wifi, made temphumidllight a struct and also made everything uints

+++ src/temperature_humidity.h
@@ -3,8 +3,8 @@
 #include <stdint.h>
 
 typedef struct{
-    float temp;
-    float humid;
+    uint8_t temp;
+    uint8_t humid;
 
 } TempHumid;
